refactor: consolidate cloud webdriver provider definitions
