+++ src/cloud-webdriver/provider-runtimes.ts
@@ -1,5 +1,3 @@
-import fs from 'node:fs';
-import path from 'node:path';
 import type {
   CloudArtifactProvider,
   CloudArtifactsQuery,
@@ -18,46 +16,38 @@ import {
   type ProviderDeviceRuntime,
   type ProviderPortReverseOptions,
 } from '../provider-device-runtime.ts';
-import { createAwsDeviceFarmWebDriverRuntime } from './aws-device-farm.ts';
-import { createBrowserStackWebDriverRuntime, uploadBrowserStackApp } from './browserstack.ts';
-import { CLOUD_WEBDRIVER_PROVIDERS, type CloudWebDriverKnownProviderName } from './providers.ts';
-import type { DefaultCloudWebDriverArtifactEnv } from './provider-registry.ts';
-import type { CloudWebDriverPlatform } from './runtime.ts';
+import {
+  CLOUD_WEBDRIVER_PROVIDER_DEFINITIONS,
+  type CloudWebDriverProviderDefinition,
+  type DefaultCloudWebDriverProviderRuntimeEnv,
+} from './provider-definitions.ts';
 
-export type DefaultCloudWebDriverProviderRuntimeEnv = DefaultCloudWebDriverArtifactEnv & {
-  BROWSERSTACK_WEBDRIVER_ENDPOINT?: string;
-  BROWSERSTACK_APP_UPLOAD_ENDPOINT?: string;
-  AGENT_DEVICE_AWS_DEVICE_FARM_PROJECT_ARN?: string;
-  AWS_DEVICE_FARM_PROJECT_ARN?: string;
-  AGENT_DEVICE_AWS_DEVICE_FARM_DEVICE_ARN?: string;
-  AWS_DEVICE_FARM_DEVICE_ARN?: string;
-  AGENT_DEVICE_AWS_DEVICE_FARM_APP_ARN?: string;
-  AWS_DEVICE_FARM_APP_ARN?: string;
-};
+export type { DefaultCloudWebDriverProviderRuntimeEnv } from './provider-definitions.ts';
 
 export function createDefaultCloudWebDriverProviderRuntimes(
   env: DefaultCloudWebDriverProviderRuntimeEnv = process.env,
 ): ProviderDeviceRuntime[] {
-  return [
-    new LazyCloudWebDriverProviderRuntime(CLOUD_WEBDRIVER_PROVIDERS.browserStack, env),
-    new LazyCloudWebDriverProviderRuntime(CLOUD_WEBDRIVER_PROVIDERS.awsDeviceFarm, env),
-  ];
+  return CLOUD_WEBDRIVER_PROVIDER_DEFINITIONS.map(
+    (definition) => new LazyCloudWebDriverProviderRuntime(definition, env),
+  );
 }
 
 class LazyCloudWebDriverProviderRuntime implements ProviderDeviceRuntime {
   readonly leaseLifecycle: LeaseLifecycleProvider;
   readonly cloudArtifacts: CloudArtifactProvider;
   readonly deviceInventoryProvider: DeviceInventoryProvider;
-  readonly provider: CloudWebDriverKnownProviderName;
+  readonly provider: CloudWebDriverProviderDefinition['provider'];
 
+  private readonly definition: CloudWebDriverProviderDefinition;
   private readonly env: DefaultCloudWebDriverProviderRuntimeEnv;
   private readonly runtimesByLeaseId = new Map<string, ProviderDeviceRuntime>();
 
   constructor(
-    provider: CloudWebDriverKnownProviderName,
+    definition: CloudWebDriverProviderDefinition,
     env: DefaultCloudWebDriverProviderRuntimeEnv,
   ) {
-    this.provider = provider;
+    this.definition = definition;
+    this.provider = definition.provider;
     this.env = env;
     this.leaseLifecycle = {
       allocate: async (lease, context) => await this.allocate(lease, context),
@@ -190,168 +180,6 @@ class LazyCloudWebDriverProviderRuntime implements ProviderDeviceRuntime {
         `${this.provider} lease allocation requires provider profile flags on the request.`,
       );
     }
-    return this.provider === CLOUD_WEBDRIVER_PROVIDERS.browserStack
-      ? await this.createBrowserStackRuntime(req, lease)
-      : this.createAwsDeviceFarmRuntime(req, lease);
+    return await this.definition.createRuntime({ req, lease, env: this.env });
   }
-
-  private async createBrowserStackRuntime(
-    req: DaemonRequest,
-    lease: DeviceLease,
-  ): Promise<ProviderDeviceRuntime> {
-    const username = requireEnv(this.env, 'BROWSERSTACK_USERNAME', 'BrowserStack');
-    const accessKey = requireEnv(this.env, 'BROWSERSTACK_ACCESS_KEY', 'BrowserStack');
-    const platform = requireRequestPlatform(req, 'BrowserStack');
-    const deviceName = requireFlag(req, 'device', 'BrowserStack requires --device <name>.');
-    const osVersion = requireFlag(
-      req,
-      'providerOsVersion',
-      'BrowserStack requires --provider-os-version <version>.',
-    );
-    const app = await resolveBrowserStackAppReference({
-      app: requireFlag(
-        req,
-        'providerApp',
-        'BrowserStack requires --provider-app <bs://app-id-or-local-path>.',
-      ),
-      cwd: req.meta?.cwd,
-      username,
-      accessKey,
-      uploadEndpoint: this.env.BROWSERSTACK_APP_UPLOAD_ENDPOINT,
-    });
-    return createBrowserStackWebDriverRuntime({
-      username,
-      accessKey,
-      platform,
-      deviceName,
-      osVersion,
-      app,
-      projectName: readFlag(req, 'providerProject'),
-      buildName: readFlag(req, 'providerBuild') ?? lease.runId,
-      sessionName: readFlag(req, 'providerSessionName') ?? lease.leaseId,
-      endpoint: this.env.BROWSERSTACK_WEBDRIVER_ENDPOINT,
-      uploadEndpoint: this.env.BROWSERSTACK_APP_UPLOAD_ENDPOINT,
-      sessionDetailsEndpoint: this.env.BROWSERSTACK_SESSION_DETAILS_ENDPOINT,
-    });
-  }
-
-  private createAwsDeviceFarmRuntime(
-    req: DaemonRequest,
-    lease: DeviceLease,
-  ): ProviderDeviceRuntime {
-    const platform = requireRequestPlatform(req, 'AWS Device Farm');
-    return createAwsDeviceFarmWebDriverRuntime({
-      projectArn: requireAwsValue(
-        req,
-        this.env,
-        'awsProjectArn',
-        'AGENT_DEVICE_AWS_DEVICE_FARM_PROJECT_ARN',
-        'AWS_DEVICE_FARM_PROJECT_ARN',
-      ),
-      deviceArn: requireAwsValue(
-        req,
-        this.env,
-        'awsDeviceArn',
-        'AGENT_DEVICE_AWS_DEVICE_FARM_DEVICE_ARN',
-        'AWS_DEVICE_FARM_DEVICE_ARN',
-      ),
-      appArn:
-        readFlag(req, 'awsAppArn') ??
-        this.env.AGENT_DEVICE_AWS_DEVICE_FARM_APP_ARN ??
-        this.env.AWS_DEVICE_FARM_APP_ARN,
-      region: readFlag(req, 'awsRegion') ?? this.env.AWS_REGION ?? this.env.AWS_DEFAULT_REGION,
-      platform,
-      deviceName: readFlag(req, 'device') ?? 'AWS Device Farm device',
-      sessionName: readFlag(req, 'providerSessionName') ?? lease.leaseId,
-      interactionMode: readAwsInteractionMode(req),
-    });
-  }
-}
-
-async function resolveBrowserStackAppReference(options: {
-  app: string;
-  cwd?: string;
-  username: string;
-  accessKey: string;
-  uploadEndpoint?: string;
-}): Promise<string> {
-  if (isProviderAppReference(options.app)) return options.app;
-  const appPath = path.resolve(options.cwd ?? process.cwd(), options.app);
-  if (!fs.existsSync(appPath)) {
-    throw new AppError(
-      'INVALID_ARGS',
-      'BrowserStack --provider-app must be a bs:// app id, URL, or existing local app path.',
-      { providerApp: options.app },
-    );
-  }
-  return await uploadBrowserStackApp(appPath, {
-    username: options.username,
-    accessKey: options.accessKey,
-    endpoint: options.uploadEndpoint,
-  });
-}
-
-function isProviderAppReference(value: string): boolean {
-  return value.startsWith('bs://') || /^https?:\/\//.test(value);
-}
-
-function requireRequestPlatform(req: DaemonRequest, providerLabel: string): CloudWebDriverPlatform {
-  const platform = req.flags?.platform;
-  if (platform === 'android' || platform === 'ios') return platform;
-  throw new AppError('INVALID_ARGS', `${providerLabel} requires --platform ios|android.`);
-}
-
-function requireFlag(
-  req: DaemonRequest,
-  key: keyof NonNullable<DaemonRequest['flags']>,
-  message: string,
-): string {
-  const value = readFlag(req, key);
-  if (value) return value;
-  throw new AppError('INVALID_ARGS', message);
-}
-
-function readFlag(
-  req: DaemonRequest,
-  key: keyof NonNullable<DaemonRequest['flags']>,
-): string | undefined {
-  const value = req.flags?.[key];
-  return typeof value === 'string' && value.length > 0 ? value : undefined;
-}
-
-function requireEnv(
-  env: DefaultCloudWebDriverProviderRuntimeEnv,
-  key: keyof DefaultCloudWebDriverProviderRuntimeEnv,
-  providerLabel: string,
-): string {
-  const value = env[key];
-  if (value) return value;
-  throw new AppError('INVALID_ARGS', `${providerLabel} requires ${key} in the environment.`);
-}
-
-function requireAwsValue(
-  req: DaemonRequest,
-  env: DefaultCloudWebDriverProviderRuntimeEnv,
-  flagKey: keyof NonNullable<DaemonRequest['flags']>,
-  primaryEnv: keyof DefaultCloudWebDriverProviderRuntimeEnv,
-  fallbackEnv: keyof DefaultCloudWebDriverProviderRuntimeEnv,
-): string {
-  const value = readFlag(req, flagKey) ?? env[primaryEnv] ?? env[fallbackEnv];
-  if (value) return value;
-  throw new AppError(
-    'INVALID_ARGS',
-    `AWS Device Farm requires --${dasherize(String(flagKey))} or ${fallbackEnv}.`,
-  );
-}
-
-function readAwsInteractionMode(
-  req: DaemonRequest,
-): 'INTERACTIVE' | 'NO_VIDEO' | 'VIDEO_ONLY' | undefined {
-  const value = readFlag(req, 'awsInteractionMode');
-  if (value === 'INTERACTIVE' || value === 'NO_VIDEO' || value === 'VIDEO_ONLY') return value;
-  return undefined;
-}
-
-function dasherize(value: string): string {
-  return value.replace(/[A-Z]/g, (match) => `-${match.toLowerCase()}`);
 }
